Many changes to process program options

+++ include/detector.hpp
@@ -118,7 +118,7 @@ namespace detection
     {
     public:
         detector(int test_id_in, std::string dest_ip, uint8_t tos, uint16_t ip_length, uint16_t id, uint16_t frag_off,
-                 uint8_t ttl, uint8_t proto, uint16_t check_sum, uint32_t sport, uint32_t dport, std::string filename,
+                 uint8_t ttl, uint8_t proto, uint16_t check_sum, uint16_t sport, uint16_t dport, std::string filename,
                  uint16_t num_packets, uint16_t data_length, uint16_t num_tail, uint16_t tail_wait,
                  raw_level raw_status, transport_type trans_proto, bool verbose_option)
             : dest_ip(dest_ip),
